10.4 compatibility
git-svn-id: https://svn.code.sf.net/p/osirix/code/osirix@999 4e9e6e3d-4551-47eb-9dbe-a6127f9400a1

+++ ROIManagerController.m
@@ -92,7 +92,7 @@ PURPOSE.
 - (IBAction)deleteROI:(id)sender
 {
 	int i;
-	NSUInteger index;
+	unsigned long index;
 	NSMutableArray* names = [NSMutableArray arrayWithCapacity:1];
 	NSIndexSet* indexSet = [tableView selectedRowIndexes];
 	index = [indexSet lastIndex];
